# rangefeed: introduce processor interface