--- a/pkg/kv/kvserver/rangefeed/bench_test.go
+++ b/pkg/kv/kvserver/rangefeed/bench_test.go
@@ -166,7 +166,7 @@ func runBenchmarkRangefeed(b *testing.B, opts benchmarkRangefeedOpts) {
 	}
 
 	// Wait for catchup scans and flush checkpoint events.
-	p.syncEventAndRegistrations()
+	syncEventAndRegistrations(p)
 
 	// Run the benchmark. We accounted for b.N when constructing events.
 	b.ResetTimer()
@@ -181,7 +181,7 @@ func runBenchmarkRangefeed(b *testing.B, opts benchmarkRangefeedOpts) {
 			b.Fatal("failed to forward closed timestamp")
 		}
 	}
-	p.syncEventAndRegistrations()
+	syncEventAndRegistrations(p)
 
 	// Check that all registrations ended successfully, and emitted the expected
 	// number of events.