This PR extracts externally visible processor functions into interface.
This will allow us to provide alternative implementations while keepingan option to fall back to old one.

Release note: None

Epic: https://cockroachlabs.atlassian.net/browse/CRDB-26372