@erikgrinaker 2023-08-07
Ran this through `BenchmarkRangefeed` just to see what the dynamic dispatch cost would be. Appears negligible, as expected:

```
name                                      old time/op  new time/op  delta
Rangefeed/opType=write/numRegs=1-24       1.72µs ± 4%  1.72µs ± 5%    ~     (p=0.922 n=18+19)
Rangefeed/opType=write/numRegs=10-24      9.25µs ± 7%  9.34µs ± 5%    ~     (p=0.428 n=20+19)
Rangefeed/opType=write/numRegs=100-24      126µs ± 1%   127µs ± 5%    ~     (p=0.078 n=18+18)
Rangefeed/opType=commit/numRegs=1-24      4.43µs ± 5%  4.38µs ± 4%    ~     (p=0.138 n=19+19)
Rangefeed/opType=commit/numRegs=10-24     11.7µs ± 7%  11.8µs ± 8%    ~     (p=0.344 n=20+20)
Rangefeed/opType=commit/numRegs=100-24     128µs ± 2%   127µs ± 2%  -0.67%  (p=0.018 n=19+17)
Rangefeed/opType=closedts/numRegs=1-24    1.39µs ± 7%  1.40µs ±10%    ~     (p=0.784 n=20+20)
Rangefeed/opType=closedts/numRegs=10-24   7.94µs ± 6%  8.10µs ± 5%  +1.97%  (p=0.009 n=19+20)
Rangefeed/opType=closedts/numRegs=100-24   118µs ± 1%   117µs ± 2%    ~     (p=0.103 n=17+18)
```